+++ test/test_helper.rb
@@ -5,3 +5,6 @@ require 'minitest/autorun'
 require 'minitest/pride'
 
 require File.expand_path('../../lib/nsrr', __FILE__)
+
+# Initialize the String class `@disable_colorization` instance variable
+String.disable_colorization = false